requirements.txt
mecab-python3
git+https://github.com/tagucci/pythonrouge.git
